backend/requirements.txt
django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
drf-yasg==1.21.7
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==4.1.0
pytest-html==4.1.1
flake8>=6.1.0
gunicorn==21.2.0
djangorestframework>=3.12.0
djangorestframework-simplejwt>=5.0.0
django-cors-headers>=3.10.0
bcrypt>=4.0.1
black>=24.3.0
isort>=5.13.2
pylint>=3.3.0
